refactor(bal): simplify module exports
