+++ crates/evm2/src/evm/bal/list.rs
@@ -1,7 +1,7 @@
 //! The top-level Block Access List structure.
 
 use super::{AccountBal, BalError, BlockAccessIndex};
-use crate::evm::state::AccountChange;
+use crate::{bytecode::BytecodeDecodeError, evm::state::AccountChange};
 use alloc::vec::Vec;
 use alloy_eip7928::{AccountChanges as AlloyAccountChanges, BlockAccessList as AlloyBal};
 use alloy_primitives::{Address, U256, map::AddressMap};
@@ -78,6 +78,54 @@ impl From<Bal> for AlloyBal {
         );
         alloy_bal.sort_unstable_by_key(|a| a.address);
         alloy_bal
+    }
+}
+
+impl TryFrom<&[AlloyAccountChanges]> for Bal {
+    type Error = BytecodeDecodeError;
+
+    /// Convert borrowed EIP-7928 [`AlloyAccountChanges`] into a [`Bal`] without consuming
+    /// the source.
+    ///
+    /// # Errors
+    ///
+    /// Returns [`BytecodeDecodeError`] if any account code change contains bytecode
+    /// rejected by [`Bytecode::new_raw_checked`](crate::bytecode::Bytecode::new_raw_checked). This
+    /// currently happens for malformed EIP-7702 bytecode, such as bytes with the EIP-7702 magic
+    /// prefix but an invalid length or unsupported version.
+    #[inline]
+    fn try_from(alloy_bal: &[AlloyAccountChanges]) -> Result<Self, Self::Error> {
+        let mut accounts =
+            AddressMap::with_capacity_and_hasher(alloy_bal.len(), Default::default());
+        for alloy_account in alloy_bal {
+            accounts.insert(alloy_account.address, AccountBal::try_from(alloy_account)?);
+        }
+
+        Ok(Self { accounts })
+    }
+}
+
+impl TryFrom<AlloyBal> for Bal {
+    type Error = BytecodeDecodeError;
+
+    /// Convert an EIP-7928 [`AlloyBal`] into a [`Bal`].
+    ///
+    /// # Errors
+    ///
+    /// Returns [`BytecodeDecodeError`] if any account code change contains bytecode
+    /// rejected by [`Bytecode::new_raw_checked`](crate::bytecode::Bytecode::new_raw_checked). This
+    /// currently happens for malformed EIP-7702 bytecode, such as bytes with the EIP-7702 magic
+    /// prefix but an invalid length or unsupported version.
+    #[inline]
+    fn try_from(alloy_bal: AlloyBal) -> Result<Self, Self::Error> {
+        let mut accounts =
+            AddressMap::with_capacity_and_hasher(alloy_bal.len(), Default::default());
+        for alloy_account in alloy_bal {
+            let address = alloy_account.address;
+            accounts.insert(address, AccountBal::try_from(alloy_account)?);
+        }
+
+        Ok(Self { accounts })
     }
 }
